Check branch protection
